# openapi/account-api.yaml
openapi: 3.0.3
info:
  title: arhDesign Account API
  version: 0.2.0
  description: R1 contract for identity, sessions, projects and super-admin user management.
  license:
    name: Proprietary
    url: https://designer-svetlana.ru/
servers:
  - url: /api
...
security:
  - accountSession: []
paths:
...
  /v1/settings:
    get:
      tags: [Settings]
      operationId: getUserSettings
      summary: Get interface settings of the current account
      responses:
        '200':
          description: Settings follow the authenticated account across browsers.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSettings'
        '401':
          $ref: '#/components/responses/Unauthenticated'
    put:
      tags: [Settings]
      operationId: updateUserSettings
      summary: Replace interface settings of the current account
      security:
        - accountSession: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserSettingsRequest'
      responses:
        '200':
          description: Settings were stored for the current account.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSettings'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/CsrfRejected'
...
  /v1/projects/{projectId}:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    get:
      tags: [Projects]
      operationId: getProject
      summary: Get a project visible to the current account
      responses:
        '200':
          description: Project visible to the actor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectView'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/ResourceHidden'
    patch:
      tags: [Projects]
      operationId: updateProject
      summary: Update a project under optimistic concurrency control
      security:
        - accountSession: []
          csrfHeader: []
      parameters:
        - $ref: '#/components/parameters/IfMatchVersion'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectRequest'
      responses:
        '200':
          description: Updated project.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectView'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ResourceHidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
    delete:
      tags: [Projects]
      operationId: deleteProject
      summary: Archive/delete a project; only super-admin may perform this R1 action
      security:
        - accountSession: []
          csrfHeader: []
      parameters:
        - $ref: '#/components/parameters/IfMatchVersion'
      responses:
        '204':
          description: Project was removed from active work; history retention still applies.
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ResourceHidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
  /v1/projects/{projectId}/settings:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    get:
      tags: [Settings]
      operationId: getProjectUserSettings
      summary: Get personal settings of the current account in one project
      responses:
        '200':
          description: Personal project settings, or documented defaults when not stored yet.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectUserSettings'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/ResourceHidden'
    put:
      tags: [Settings]
      operationId: updateProjectUserSettings
      summary: Replace personal settings of the current account in one project
      security:
        - accountSession: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectUserSettingsRequest'
      responses:
        '200':
          description: Personal project settings were stored.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectUserSettings'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/CsrfRejected'
        '404':
          $ref: '#/components/responses/ResourceHidden'
...
components:
  securitySchemes:
    accountSession:
      type: apiKey
      in: cookie
      name: __Host-arhdesign_session
      description: Opaque server-side session cookie. Raw values are never exposed to JavaScript.
    refreshSession:
      type: apiKey
      in: cookie
      name: __Secure-arhdesign_refresh
      description: Path-scoped, single-use refresh cookie. Raw values are never exposed to JavaScript.
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Synchronizer token matching the readable CSRF cookie on state-changing requests.
    telegramWebhookSecret:
      type: apiKey
      in: header
      name: X-Telegram-Bot-Api-Secret-Token
      description: Constant-time compared secret configured when the Telegram webhook is registered.
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    IfMatchVersion:
      name: If-Match
      in: header
      required: true
      description: Quoted decimal optimistic version returned by the previous representation.
      schema:
        type: string
        pattern: '^"[1-9][0-9]*"$'
  responses:
    ValidationError:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    RateLimited:
      description: Request rate limit was reached.
      headers:
        Retry-After:
          required: true
          schema:
            type: integer
            minimum: 1
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    Unauthenticated:
      description: No active authenticated session.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    CsrfRejected:
      description: Origin or CSRF token validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Forbidden:
      description: Actor is authenticated but not allowed to perform this action.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    ResourceHidden:
      description: Resource does not exist or is hidden from this actor.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Conflict:
      description: Optimistic version or domain uniqueness conflict.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    ServiceUnavailable:
      description: A required security dependency is temporarily unavailable; the operation did not execute.
      headers:
        Retry-After:
          required: true
          schema:
            type: integer
            minimum: 1
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
    ThemePreference:
      type: string
      enum: [light, dark]
    UserSettings:
      type: object
      additionalProperties: false
      required: [theme]
      properties:
        theme:
          $ref: '#/components/schemas/ThemePreference'
    UpdateUserSettingsRequest:
      type: object
      additionalProperties: false
      required: [theme]
      properties:
        theme:
          $ref: '#/components/schemas/ThemePreference'
    ProjectUserSettings:
      type: object
      additionalProperties: false
      required: [upcomingDays]
      properties:
        upcomingDays:
          type: integer
          minimum: 1
          maximum: 90
          default: 7
    UpdateProjectUserSettingsRequest:
      type: object
      additionalProperties: false
      required: [upcomingDays]
      properties:
        upcomingDays:
          type: integer
          minimum: 1
          maximum: 90
...
    ProjectStatus:
      type: string
      enum: [draft, active, paused, completed, pending_deletion, archived]
    ProjectView:
      type: object
      additionalProperties: false
      required: [id, name, type, status, currencyCode, autoApproveExpenses, createdByUserId, createdAt, version]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          type: string
        address:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/ProjectStatus'
        customerUserId:
          type: string
          format: uuid
          nullable: true
        adminUserIds:
          type: array
          maxItems: 50
          items:
            type: string
            format: uuid
        plannedStartOn:
          type: string
          format: date
          nullable: true
        plannedFinishOn:
          type: string
          format: date
          nullable: true
        currencyCode:
          type: string
          pattern: '^[A-Z]{3}$'
        description:
          type: string
          nullable: true
        autoApproveExpenses:
          type: boolean
        createdByUserId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        version:
          type: integer
          format: int64
          minimum: 1
...
    UpdateProjectRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 200
        type:
          type: string
          minLength: 2
          maxLength: 100
        address:
          type: string
          maxLength: 500
          nullable: true
        status:
          $ref: '#/components/schemas/ProjectStatus'
        plannedStartOn:
          type: string
          format: date
          nullable: true
        plannedFinishOn:
          type: string
          format: date
          nullable: true
        description:
          type: string
          maxLength: 5000
          nullable: true
        autoApproveExpenses:
          type: boolean
...
    ApiError:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, requestId]
          properties:
            code:
              type: string
              pattern: '^[a-z][a-z0-9_]{2,63}$'
            message:
              type: string
              minLength: 1
              maxLength: 200
            requestId:
              type: string
              minLength: 8
              maxLength: 100
            fields:
              type: object
              maxProperties: 20
              additionalProperties:
                type: string
                maxLength: 200
              description: Field messages never repeat submitted passwords, tokens or full PII.